Move KeyEvent::encode in binding module
